src/python/py-authorizer/requirements.txt
aws-lambda-typing==2.19.0
cffi==1.16.0
cryptography==42.0.5
pycparser==2.21
PyJWT==2.8.0
